# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lオペ for CLINIC API",
    "description": "クリニック特化LINE運用プラットフォーム — API仕様書",
    "version": "1.0.0",
    "contact": {
      "name": "Lオペ for CLINIC"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/admin/doctors": {
      "get": {
        "summary": "医師一覧取得",
        "tags": [
          "管理: スケジュール"
        ],
        "responses": {
          "200": {
            "description": "医師一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "adminCookie": []
          }
        ]
      },
      "post": {
        "summary": "医師登録・更新",
        "tags": [
          "管理: スケジュール"
        ],
        "responses": {
          "200": {
            "description": "登録結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "adminCookie": []
          }
        ]
      }
    },
...
    "/api/admin/karte-templates": {
      "get": {
        "summary": "カルテテンプレート一覧",
        "tags": [
          "管理: カルテ"
        ],
        "responses": {
          "200": {
            "description": "テンプレート一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "adminCookie": []
          }
        ]
      },
      "post": {
        "summary": "カルテテンプレート作成",
        "tags": [
          "管理: カルテ"
        ],
        "responses": {
          "200": {
            "description": "作成結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "adminCookie": []
          }
        ]
      }
    },
...
    "/api/admin/line/rich-menus": {
      "get": {
        "summary": "リッチメニュー一覧",
        "tags": [
          "管理: LINE"
        ],
        "responses": {
          "200": {
            "description": "リッチメニュー一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "adminCookie": []
          }
        ]
      },
      "post": {
        "summary": "リッチメニュー作成",
        "tags": [
          "管理: LINE"
        ],
        "responses": {
          "200": {
            "description": "作成結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー"
          },
          "500": {
            "description": "サーバーエラー"
          }
        },
        "security": [
          {
            "adminCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "chat_bar_text": {
                    "type": "string"
                  },
                  "size_type": {
                    "type": "string"
                  },
                  "areas": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "selected": {
                    "type": "boolean"
                  },
                  "image_url": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session",
        "description": "管理画面セッションCookie"
      },
      "patientCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "patient_id",
        "description": "患者IDCookie（LINE認証後に自動付与）"
      },
      "platformCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "platform_session",
        "description": "プラットフォーム管理セッションCookie"
      }
    }
  },
...
}